WebApp/requirements.txt
flask==2.1.1
databricks-sql-connector==1.0.2
